backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Cors==5.0.0
python-dotenv==1.0.1
Werkzeug==3.1.3
requests==2.32.3
